rename local_semaphore to semaphore
