+++ scipio/src/lib.rs
@@ -131,18 +131,18 @@ pub mod collections;
 mod dma_file;
 mod error;
 mod executor;
-mod local_semaphore;
 mod multitask;
 mod networking;
 mod pollable;
+mod semaphore;
 pub mod timer;
 
 pub use crate::dma_file::{Directory, DmaFile};
 pub use crate::error::Error;
 pub use crate::executor::{LocalExecutor, QueueNotFoundError, Task, TaskQueueHandle};
-pub use crate::local_semaphore::Semaphore;
 pub use crate::networking::*;
 pub use crate::pollable::Async;
+pub use crate::semaphore::Semaphore;
 pub use crate::sys::DmaBuffer;
 pub use crate::timer::{Timer, TimerActionOnce, TimerActionRepeat};
 pub use enclose::enclose;
